+++ src/components/create-site-header-trip-cost.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createCostTripTemplate = () => {
   return (
@@ -8,23 +8,10 @@ const createCostTripTemplate = () => {
   );
 };
 
-export default class CostTrip {
-  constructor() {
-    this._element = null;
-  }
+export default class CostTrip extends AbstractComponent {
 
   getTemplate() {
     return createCostTripTemplate();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
