+++ src/components/create-site-header-trip-info.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createInfoTripTemplate = () => {
   return (
@@ -10,23 +10,10 @@ const createInfoTripTemplate = () => {
   );
 };
 
-export default class InfoContainer {
-  constructor() {
-    this._element = null;
-  }
+export default class InfoContainer extends AbstractComponent {
 
   getTemplate() {
     return createInfoTripTemplate();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
