+++ src/components/create-site-maintContent-filter-sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createSortTripTemplate = () => {
   return (
@@ -35,23 +35,10 @@ const createSortTripTemplate = () => {
   );
 };
 
-export default class SortTrip {
-  constructor() {
-    this._element = null;
-  }
+export default class SortTrip extends AbstractComponent {
 
   getTemplate() {
     return createSortTripTemplate();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
