+++ src/components/create-site-maintContent-listDay.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createTripDaysList = () => {
   return (
@@ -8,23 +8,10 @@ const createTripDaysList = () => {
   );
 };
 
-export default class TripDaysList {
-  constructor() {
-    this._element = null;
-  }
+export default class TripDaysList extends AbstractComponent {
 
   getTemplate() {
     return createTripDaysList();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
