+++ src/components/create-site-maintContent-no-points.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const createNoPointsTemplate = () => {
   return (
@@ -8,24 +8,11 @@ const createNoPointsTemplate = () => {
   );
 };
 
-export default class NoPoints {
-  constructor() {
-    this._element = null;
-  }
+export default class NoPoints extends AbstractComponent {
 
   getTemplate() {
     return createNoPointsTemplate();
   }
 
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
 }
