+++ src/utils.js
@@ -1,82 +1,59 @@
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
+// const castTimeFormat = (value) => {
+//   return value < 10 ? `0${value}` : String(value);
+// };
 
-  return newElement.firstChild;
-};
+// const castTimeMinutesFormat = (value) => {
+//   return value < 30 ? `00` : `30`;
+// };
 
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-};
+// const formatTime = (date) => {
 
-export const renderTemplate = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
+//   const hours = castTimeFormat(date.getHours() % 24);
+//   const minutes = castTimeMinutesFormat(date.getMinutes());
 
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
+//   return `${hours}:${minutes}`;
+// };
 
-const castTimeMinutesFormat = (value) => {
-  return value < 30 ? `00` : `30`;
-};
+// const formatDate = (date) => {
+//   const dateNow = castTimeFormat(date.getDate());
+//   const month = castTimeFormat(date.getMonth() + 1);
+//   const year = castTimeFormat(date.getFullYear()).slice(2);
 
-const formatTime = (date) => {
+//   const fullDate = `${dateNow}/${month}/${year}`;
 
-  const hours = castTimeFormat(date.getHours() % 24);
-  const minutes = castTimeMinutesFormat(date.getMinutes());
+//   return fullDate;
+// };
 
-  return `${hours}:${minutes}`;
-};
+// const getRandomArrayItem = (array) => {
+//   const randomIndex = getRandomIntegerNumber(0, array.length);
+//   return array[randomIndex];
+// };
 
-const formatDate = (date) => {
-  const dateNow = castTimeFormat(date.getDate());
-  const month = castTimeFormat(date.getMonth() + 1);
-  const year = castTimeFormat(date.getFullYear()).slice(2);
+// const getRandomIntegerNumber = (min, max) => {
+//   return min + Math.floor(Math.random() * (max - min));
+// };
 
-  const fullDate = `${dateNow}/${month}/${year}`;
+// function shuffleArray(array) {
+//   const result = array.slice();
+//   result.sort(() => Math.random() - 0.5);
+//   return result;
+// }
 
-  return fullDate;
-};
+// const getStartDate = () => {
+//   const targetDate = new Date();
+//   return targetDate;
+// };
 
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomIntegerNumber(0, array.length);
-  return array[randomIndex];
-};
+// const getRandomDate = () => {
+//   const targetDate = new Date();
+//   const diffValueDate = getRandomIntegerNumber(2, 15);
+//   const diffValueHours = getRandomIntegerNumber(0, 24);
+//   const diffValueMinutes = getRandomIntegerNumber(0, 60);
+//   targetDate.setDate(targetDate.getDate() + diffValueDate);
+//   targetDate.setHours(targetDate.getHours() + diffValueHours);
+//   targetDate.setMinutes(targetDate.getMinutes() + diffValueMinutes);
 
-const getRandomIntegerNumber = (min, max) => {
-  return min + Math.floor(Math.random() * (max - min));
-};
+//   return targetDate;
+// };
 
-function shuffleArray(array) {
-  const result = array.slice();
-  result.sort(() => Math.random() - 0.5);
-  return result;
-}
-
-const getStartDate = () => {
-  const targetDate = new Date();
-  return targetDate;
-};
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const diffValueDate = getRandomIntegerNumber(2, 15);
-  const diffValueHours = getRandomIntegerNumber(0, 24);
-  const diffValueMinutes = getRandomIntegerNumber(0, 60);
-  targetDate.setDate(targetDate.getDate() + diffValueDate);
-  targetDate.setHours(targetDate.getHours() + diffValueHours);
-  targetDate.setMinutes(targetDate.getMinutes() + diffValueMinutes);
-
-  return targetDate;
-};
-
-export {formatTime, formatDate, getRandomArrayItem, getRandomIntegerNumber, shuffleArray, getStartDate, getRandomDate};
+// export {formatTime, formatDate, getRandomArrayItem, getRandomIntegerNumber, shuffleArray, getStartDate, getRandomDate};
